refactor: CategorySection, SelectFriendSection, MainPage style 분리 및 theme 토큰 적용
