+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { Global } from "@emotion/react";
 import reset from "@/styles/reset";
 import Layout from "@/layout/Layout";
-import MainPage from "@/pages/MainPage";
+import MainPage from "@/pages/MainPage/MainPage";
 
 function App() {
   return (
